+++ carp_context_package/ios/Flutter/flutter_export_environment.sh
@@ -1,0 +1,11 @@
+#!/bin/sh
+# This is a generated file; do not edit or check into version control.
+export "FLUTTER_ROOT=/Users/bardram/dev/flutter"
+export "FLUTTER_APPLICATION_PATH=/Users/bardram/dev/carp/carp.sensing-flutter/carp_context_package"
+export "FLUTTER_TARGET=lib/main.dart"
+export "FLUTTER_BUILD_DIR=build"
+export "SYMROOT=${SOURCE_ROOT}/../build/ios"
+export "OTHER_LDFLAGS=$(inherited) -framework Flutter"
+export "FLUTTER_FRAMEWORK_DIR=/Users/bardram/dev/flutter/bin/cache/artifacts/engine/ios"
+export "FLUTTER_BUILD_NAME=0.6.6"
+export "FLUTTER_BUILD_NUMBER=0.6.6"